Rename plugin in code
This plugin has gone far beyond the initial hello world thing that
it was originally named for. Rename it to MDNS to better reflect
what it is doing now.

Just the names in the code are changed in this commit. I'll rename
the main plugin file in a followup so diffs can still be sanely viewed.