Move mdns browsing to background goroutine
Instead of browsing mdns for every query, which takes around 2
seconds every time, run the browsing in a background goroutine that
populates the hosts maps for the plugin to use immediately.

The background browsing happens every 30 seconds because that is the
configured caching time, so in general results will be cached for
that long anyway.