Shorten TTL and background loop delay
With 60 second TTLs, it was possible for entries to get cached for
up to 30 seconds (the default OpenShift setting for the cache plugin).
Since our entries are dynamic, we don't want to wait 30 seconds
every time the environment changes. This shortens our TTLs to 0
and the loop delay to 5 seconds, which seems to be the minimum time
that the cache plugin will cache. Since it takes around 2 seconds
to browse the mDNS records anyway, this leaves a fairly short period
of time where our entries might be inconsistent.

Note that this doesn't address A record caching because those
shouldn't change after a node starts broadcasting.