Fix improper .local addresses when use-ipv6 enabled
It seems that when ipv6 addresses are also being broadcast, the mdns
library may update our entry objects after we modify them. This
overwrites our changes and results in unmodified .local addresses
being returned to the client.

To prevent this, make a local copy of the entry while processing it.
This copy can't be modified by mdns later so it eliminates the problem.

Note that even with this change, when use-ipv6 is enabled it is
necessary to also set:

publish-a-on-ipv6=yes

Otherwise we get mdns entries with no ipv4 address set, which the
plugin currently can't handle.