Fix mutex handling
It appears that copying a mutex actually creates a new mutex that
doesn't coordinate with any other copies of the mutex. Because the
plugin architecture uses a value receiver, that means we need to make
the mutex a pointer instead of inheriting it in the plugin. Otherwise
every call to ServeDNS gets a different mutex and no meaningful
locking occurs.