Lock only long enough to update host maps
Instead of locking the mutex for the entire duration of the mdns
browse process, do the browsing into local variables that don't need
locking and then lock just long enough to copy those values into the
shared structures. This prevents potential long delays if a query
happens to come in while we're updating.