Extract cname index from hostname
Instead of generating cname indices from the loop counter, extract
them from the target hostname. This prevents things like having
etcd-1 point at master-2 if master-0 or master-1 haven't come up yet.

This does add a requirement on the master hostnames that they must
end with "-#" so we know where to look for the index.