Don't add nil address records
Up until now the plugin assumed that there was always both an IPv4
and IPv6 address present in the mdns entries. That isn't always true
and it results in malformed responses if one of them is nil. We
should only populate the address records that we actually have data
for.