Pre-populate bootstrap SRV and CNAME records
On initial deployment, if one node comes up before the others it may
get an incomplete SRV record list from the plugin. We need to prevent
that as it breaks clustering once the other nodes come up. This
change adds static SRV and CNAME records for master-{0-2} so at
minimum those records will always be present. The hosts they point
at may not be available right away, but that's the same way it
worked with static records before.