Add ability to filter mdns records by name
Adds another configuration value to the plugin that can be used to
filter the records received from mdns-publisher. This allows us to
include the cluster name in the service names and avoid reading
records from any other mdns services that may be running on the same
network.

The filter uses a simple substring match, so the value used should
be as unique as possible to avoid accidental matches.