Switch to grandcat/zeroconf
With the whyrusleeping mdns library we were having issues when there
were large numbers of services. Sometimes we weren't getting a full
list back when we queried for them. Avahi found them successfully, so
the problem seems to have been on the client side.

Switching to grandcat/zeroconf for our client library seems to have
fixed this problem. It also has some handy features we might take
advantage of in the future, like supporting multiple IPs per record.