Add Makefile for ci
Like mdns-publisher, we want to have a makefile in this repo that
we can use to run things like go fmt and go vet.

Note that at this time go vet is failing for me on this repo. I
don't think it's a valid complaint, but I'm not sure how to fix it
either.