Actually use local source in build-coredns
It turns out the build script didn't use the local copy of coredns-mdns
because it is vendored in openshift/coredns. This change copies the
local source into the vendor directory so changes will be properly
included.

Because we are no longer appending to a file in coredns, it is also
not necessary to delete coredns before every build. Skipping the
git clone when reusing a build directory speeds things up immensely.