Add support for building coredns image in build-coredns.sh
For e2e testing purposes we need to be able to do this. This reuses
the logic in build-coredns.sh to build the image using the local
changes in coredns-mdns. The only difference is it runs a container
build instead of just building the binary.

To enable the image build, export CONTAINER_IMAGE with a value that
refers to an image repo. Because this also depends on the Dockerfile
in the openshift fork of coredns, COREDNS_REPO must also be set
appropriately. For example:

export CONTAINER_IMAGE=quay.io/cybertron/coredns
export COREDNS_REPO=https://github.com/openshift/coredns