Add bind address configuration setting
Previously coredns-mdns sent multicast traffic to all available
interfaces on the system. This is problematic in some environments
where it is undesirable to send such traffic to some interfaces.
For example, tunnel interfaces where a specific encapsulation is
required to avoid errors.

This change adds a new configuration option to the plugin. When
specified, the plugin will look up the interface associated with
the address and only send multicast traffic on that interface. If
the option is not specified, the behavior will be as before.