Vendor deps
For some reason when I try to update the unit test job to use a more
recent version of Go (which needs to happen due to some changes in
the openshift fork of coredns), it insists on looking for deps in the
vendor directory. Hopefully this will unblock the repo and allow us to
move forward.