Update caddy import to use the one from coredns
When caddy moved to v2, coredns chose to stay with v1 by moving it
in-tree. Plugins need to update their imports to reflect this.

Note that this will fail the build-coredns ci job until the
openshift/coredns repo has been updated to the new version. I have
tested it locally with a newer version and it works.